refactor: extract code generation into module
